feat: Update Header component to fetch form submissions from mock server
This commit updates the Header component by importing the getFormSubmissions function from the ./service/mockServer module. The createMockFormSubmission function that was previously used to add new submissions has been removed from the onClick event of the New Submission button.

Instead, a call to getFormSubmissions is made on click of the New Submission button to update the list of form submissions. This ensures that the form submissions list is always up-to-date.

This change was made to improve the user experience by ensuring that users have access to the latest form submissions when they click the New Submission button.