Refactor App.js to use arrow function and div container
In this commit, I refactored the App.js file to use arrow function syntax and wrapped the Header and Content components in a div container instead of a Fragment. I also changed the way the useEffect hook is defined to use an arrow function syntax. This change improves the readability and maintainability of the code.