Refactor Content component to use Material UI Typography and destructured props
We refactored the Content component to use Material UI Typography instead of MUI Box and Typography components. We also destructured the props passed to the component to make the code cleaner and more concise. This change simplifies the code and makes it more readable while still achieving the same functionality.