Refactor App component to pass down setLikedSubmissions function to Header component
The initial App component made an HTTP request to fetch the liked submissions data using the useEffect hook. However, this logic has been moved to the Content component, which is a better place for it.

To simplify the App component, I removed the useEffect hook and instead passed down the setLikedSubmissions function to the Header component. This way, the Header component can trigger a re-fetch of the liked submissions data if needed.